[Common] Refactor the enum headers for supporting both C and C++.
