+++ include/CMakeLists.txt
@@ -32,6 +32,9 @@ CHECK_INCLUDE_FILE_CXX(pwd.h HAVE_PWD_H)
 configure_file(common/config.h.in common/config.h)
 configure_file(common/version.h.in common/version.h)
 configure_file(api/wasmedge/version.h.in api/wasmedge/version.h)
+configure_file(common/enum_configure.h api/wasmedge/enum_configure.h COPYONLY)
+configure_file(common/enum_errcode.h api/wasmedge/enum_errcode.h COPYONLY)
+configure_file(common/enum_types.h api/wasmedge/enum_types.h COPYONLY)
 unset(WASMEDGE_VERSION_STRING)
 unset(WASMEDGE_VERSION_LIST)
 unset(WASMEDGE_VERSION_MAJOR)
