+++ include/api/wasmedge/wasmedge.h
@@ -29,190 +29,11 @@
 #include <stdint.h>
 #include <stdlib.h>
 
+#include "wasmedge/enum_configure.h"
+#include "wasmedge/enum_errcode.h"
+#include "wasmedge/enum_types.h"
 #include "wasmedge/int128.h"
 #include "wasmedge/version.h"
-
-/// Error code enumeration.
-enum WasmEdge_ErrCode {
-  WasmEdge_ErrCode_Success = 0x00,
-  /// Exit and return success.
-  WasmEdge_ErrCode_Terminated = 0x01,
-  /// Generic runtime error.
-  WasmEdge_ErrCode_RuntimeError = 0x02,
-  /// Exceeded cost limit (out of gas).
-  WasmEdge_ErrCode_CostLimitExceeded = 0x03,
-  /// Wrong VM's workflow
-  WasmEdge_ErrCode_WrongVMWorkflow = 0x04,
-  /// Wasm function not found
-  WasmEdge_ErrCode_FuncNotFound = 0x05,
-  /// AOT runtime is disabled
-  WasmEdge_ErrCode_AOTDisabled = 0x06,
-
-  /// Load phase
-  /// File not found
-  WasmEdge_ErrCode_InvalidPath = 0x20,
-  /// Error when reading
-  WasmEdge_ErrCode_ReadError = 0x21,
-  /// Reach end of file when reading
-  WasmEdge_ErrCode_UnexpectedEnd = 0x22,
-  /// Not detected magic header
-  WasmEdge_ErrCode_InvalidMagic = 0x23,
-  /// Unsupported version
-  WasmEdge_ErrCode_InvalidVersion = 0x24,
-  /// Malformed section ID
-  WasmEdge_ErrCode_InvalidSection = 0x25,
-  /// Section size mismatched
-  WasmEdge_ErrCode_SectionSizeMismatch = 0x26,
-  /// Name size out of bounds
-  WasmEdge_ErrCode_NameSizeOutOfBounds = 0x27,
-  /// Junk sections
-  WasmEdge_ErrCode_JunkSection = 0x28,
-  /// Incompatible function and code section
-  WasmEdge_ErrCode_IncompatibleFuncCode = 0x29,
-  /// Incompatible data and datacount section
-  WasmEdge_ErrCode_IncompatibleDataCount = 0x2A,
-  /// Datacount section required
-  WasmEdge_ErrCode_DataCountRequired = 0x2B,
-  /// Malformed import kind
-  WasmEdge_ErrCode_InvalidImportKind = 0x2C,
-  /// Malformed export kind
-  WasmEdge_ErrCode_InvalidExportKind = 0x2D,
-  /// Not loaded an expected zero byte
-  WasmEdge_ErrCode_ExpectedZeroByte = 0x2E,
-  /// Malformed mutability
-  WasmEdge_ErrCode_InvalidMut = 0x2F,
-  /// Local size too large
-  WasmEdge_ErrCode_TooManyLocals = 0x30,
-  /// Malformed value type
-  WasmEdge_ErrCode_InvalidValType = 0x31,
-  /// Malformed element type (Bulk-mem proposal)
-  WasmEdge_ErrCode_InvalidElemType = 0x32,
-  /// Malformed reference type (Ref-types proposal)
-  WasmEdge_ErrCode_InvalidRefType = 0x33,
-  /// Invalid utf-8 encoding
-  WasmEdge_ErrCode_InvalidUTF8 = 0x34,
-  /// Invalid too large integer
-  WasmEdge_ErrCode_IntegerTooLarge = 0x35,
-  /// Invalid presentation too long integer
-  WasmEdge_ErrCode_IntegerTooLong = 0x36,
-  /// Illegal OpCode
-  WasmEdge_ErrCode_InvalidOpCode = 0x37,
-  /// Parsing error
-  WasmEdge_ErrCode_InvalidGrammar = 0x38,
-
-  /// Validation phase
-  /// Alignment > natural
-  WasmEdge_ErrCode_InvalidAlignment = 0x40,
-  /// Got unexpected type when checking
-  WasmEdge_ErrCode_TypeCheckFailed = 0x41,
-  /// Branch to unknown label index
-  WasmEdge_ErrCode_InvalidLabelIdx = 0x42,
-  /// Access unknown local index
-  WasmEdge_ErrCode_InvalidLocalIdx = 0x43,
-  /// Type index not defined
-  WasmEdge_ErrCode_InvalidFuncTypeIdx = 0x44,
-  /// Function index not defined
-  WasmEdge_ErrCode_InvalidFuncIdx = 0x45,
-  /// Table index not defined
-  WasmEdge_ErrCode_InvalidTableIdx = 0x46,
-  /// Memory index not defined
-  WasmEdge_ErrCode_InvalidMemoryIdx = 0x47,
-  /// Global index not defined
-  WasmEdge_ErrCode_InvalidGlobalIdx = 0x48,
-  /// Element segment index not defined
-  WasmEdge_ErrCode_InvalidElemIdx = 0x49,
-  /// Data segment index not defined
-  WasmEdge_ErrCode_InvalidDataIdx = 0x4A,
-  /// Undeclared reference
-  WasmEdge_ErrCode_InvalidRefIdx = 0x4B,
-  /// Should be constant expression
-  WasmEdge_ErrCode_ConstExprRequired = 0x4C,
-  /// Export name conflicted
-  WasmEdge_ErrCode_DupExportName = 0x4D,
-  /// Tried to store to const global value
-  WasmEdge_ErrCode_ImmutableGlobal = 0x4E,
-  /// Invalid result arity in select t* instruction
-  WasmEdge_ErrCode_InvalidResultArity = 0x4F,
-  /// #Tables > 1 (without Ref-types proposal)
-  WasmEdge_ErrCode_MultiTables = 0x50,
-  /// #Memories > 1
-  WasmEdge_ErrCode_MultiMemories = 0x51,
-  /// Invalid Limit grammar
-  WasmEdge_ErrCode_InvalidLimit = 0x52,
-  /// Memory pages > 65536
-  WasmEdge_ErrCode_InvalidMemPages = 0x53,
-  /// Invalid start function signature
-  WasmEdge_ErrCode_InvalidStartFunc = 0x54,
-  /// Invalid lane index
-  WasmEdge_ErrCode_InvalidLaneIdx = 0x55,
-
-  /// Instantiation phase
-  /// Module name conflicted when importing.
-  WasmEdge_ErrCode_ModuleNameConflict = 0x60,
-  /// Import matching failed
-  WasmEdge_ErrCode_IncompatibleImportType = 0x61,
-  /// Unknown import instances
-  WasmEdge_ErrCode_UnknownImport = 0x62,
-  /// Init failed when instantiating data segment
-  WasmEdge_ErrCode_DataSegDoesNotFit = 0x63,
-  /// Init failed when instantiating element segment
-  WasmEdge_ErrCode_ElemSegDoesNotFit = 0x64,
-
-  /// Execution phase
-  /// Wrong access of instances addresses
-  WasmEdge_ErrCode_WrongInstanceAddress = 0x80,
-  /// Wrong access of instances indices
-  WasmEdge_ErrCode_WrongInstanceIndex = 0x81,
-  /// Instruction type not match
-  WasmEdge_ErrCode_InstrTypeMismatch = 0x82,
-  /// Function signature not match when invoking
-  WasmEdge_ErrCode_FuncSigMismatch = 0x83,
-  /// Divide by zero
-  WasmEdge_ErrCode_DivideByZero = 0x84,
-  /// Integer overflow
-  WasmEdge_ErrCode_IntegerOverflow = 0x85,
-  /// Cannot do convert to integer
-  WasmEdge_ErrCode_InvalidConvToInt = 0x86,
-  /// Out of bounds table access
-  WasmEdge_ErrCode_TableOutOfBounds = 0x87,
-  /// Out of bounds memory access
-  WasmEdge_ErrCode_MemoryOutOfBounds = 0x88,
-  /// Meet an unreachable instruction
-  WasmEdge_ErrCode_Unreachable = 0x89,
-  /// Uninitialized element in table instance
-  WasmEdge_ErrCode_UninitializedElement = 0x8A,
-  /// Access undefined element in table instances
-  WasmEdge_ErrCode_UndefinedElement = 0x8B,
-  /// Func type mismatch in call_indirect
-  WasmEdge_ErrCode_IndirectCallTypeMismatch = 0x8C,
-  /// Host function execution failed
-  WasmEdge_ErrCode_ExecutionFailed = 0x8D,
-  /// Reference type not match
-  WasmEdge_ErrCode_RefTypeMismatch = 0x8E
-};
-
-/// WASM Value type enumeration.
-enum WasmEdge_ValType {
-  WasmEdge_ValType_I32 = 0x7FU,
-  WasmEdge_ValType_I64 = 0x7EU,
-  WasmEdge_ValType_F32 = 0x7DU,
-  WasmEdge_ValType_F64 = 0x7CU,
-  WasmEdge_ValType_V128 = 0x7BU,
-  WasmEdge_ValType_FuncRef = 0x70U,
-  WasmEdge_ValType_ExternRef = 0x6FU
-};
-
-/// WASM Reference type enumeration.
-enum WasmEdge_RefType {
-  WasmEdge_RefType_FuncRef = 0x70U,
-  WasmEdge_RefType_ExternRef = 0x6FU
-};
-
-/// WASM Mutability enumeration.
-enum WasmEdge_Mutability {
-  WasmEdge_Mutability_Const = 0x00U,
-  WasmEdge_Mutability_Var = 0x01U
-};
 
 /// WasmEdge WASM value struct.
 typedef struct WasmEdge_Value {
@@ -249,49 +70,6 @@ typedef struct WasmEdge_Limit {
   /// Maximum value. Will be ignored if the `HasMax` is false.
   uint32_t Max;
 } WasmEdge_Limit;
-
-/// WASM Proposal enumeration.
-/// This enum is copied from "include/common/configure.h" and should be the
-/// same.
-enum WasmEdge_Proposal {
-  WasmEdge_Proposal_ImportExportMutGlobals = 0,
-  WasmEdge_Proposal_NonTrapFloatToIntConversions,
-  WasmEdge_Proposal_SignExtensionOperators,
-  WasmEdge_Proposal_MultiValue,
-  WasmEdge_Proposal_BulkMemoryOperations,
-  WasmEdge_Proposal_ReferenceTypes,
-  WasmEdge_Proposal_SIMD,
-  WasmEdge_Proposal_TailCall,
-  WasmEdge_Proposal_Annotations,
-  WasmEdge_Proposal_Memory64,
-  WasmEdge_Proposal_Threads,
-  WasmEdge_Proposal_ExceptionHandling,
-  WasmEdge_Proposal_FunctionReferences
-};
-
-/// Host Module Registration enumeration.
-enum WasmEdge_HostRegistration {
-  WasmEdge_HostRegistration_Wasi = 0,
-  WasmEdge_HostRegistration_WasmEdge_Process
-};
-
-/// AOT compiler optimization level enumeration.
-enum WasmEdge_CompilerOptimizationLevel {
-  /// Disable as many optimizations as possible.
-  WasmEdge_CompilerOptimizationLevel_O0 = 0,
-  /// Optimize quickly without destroying debuggability.
-  WasmEdge_CompilerOptimizationLevel_O1,
-  /// Optimize for fast execution as much as possible without triggering
-  /// significant incremental compile time or code size growth.
-  WasmEdge_CompilerOptimizationLevel_O2,
-  /// Optimize for fast execution as much as possible.
-  WasmEdge_CompilerOptimizationLevel_O3,
-  /// Optimize for small code size as much as possible without triggering
-  /// significant incremental compile time or execution time slowdowns.
-  WasmEdge_CompilerOptimizationLevel_Os,
-  /// Optimize for small code size as much as possible.
-  WasmEdge_CompilerOptimizationLevel_Oz
-};
 
 /// Opaque struct of WasmEdge configure.
 typedef struct WasmEdge_ConfigureContext WasmEdge_ConfigureContext;
