+++ include/ast/instruction.h
@@ -15,7 +15,8 @@
 #include <memory>
 #include <vector>
 
-#include "common/astdef.h"
+#include "common/configure.h"
+#include "common/enum_ast.h"
 #include "common/errcode.h"
 #include "common/types.h"
 #include "common/value.h"
