+++ lib/vm/CMakeLists.txt
@@ -7,6 +7,7 @@ wasmedge_add_library(wasmedgeVM
 target_link_libraries(wasmedgeVM
   PUBLIC
   wasmedgeCommon
+  wasmedgeSystem
   wasmedgeAST
   wasmedgeLoader
   wasmedgeValidator
